apis/requirements.txt
fastapi
sqlalchemy
pymysql
alembic
uvicorn
passlib
databases
aiomysql
python-jose